pyproject.toml
[tool.poetry]
name = "vk-scripts"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "vk_scripts", from="src"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
pyyaml = "^6.0"
aiohttp = "^3.8.4"
click = "^8.1.3"
openpyxl = "^3.1.2"


[tool.poetry.scripts]
vk-scripts = "vk_scripts.main:vk_scripts_cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
